Move scanners/fujitsu.ts to fujitsu_scanner.ts
Also inline scanners/types.ts in that file
