--- services/scan/src/scanners/index.ts
@@ -1,2 +1,0 @@
-export * from './fujitsu';
-export * from './types';
